Nightly commit :moon:
SGSM is in working order (nearly), the remaining bit of logic is to not
make the engine crash when a sprite leaves the screen bounds. That
should be the work of the yet-to-be-created, yet-to-be-designed, and
yet-to-be-thought-about GameManager.
In other news, AABBs are better. Sprites have a way to deal with sprite
collisions, now that SGSM actually registers them.