Late nightly commit :stars:
Note: Commit message reads "late nightly commit" not "late night
commit," as in I'm SUPER BEHIND on trying to be able to call it a
"nightly" commit.
This sums up a few days of work, rewriting the old engine from scratch
in Java and making a few improvements along the way (mainly to make the
concepts implemented more friendly to Java).