Big collisions update
Almost enough to call it a platformer engine! Collisions are now easier
to maintain via AABB#type(), which returns a class. This can be used,
for instance, in SolidSpriteAABB#collisions to easily maintain the map
of collision information.
More default Collision implementations exist: specifically platform and
slope (which required a new type of AABB to use).
The TileGridMap class now actually works as intended, and won't crash
the program when a sprite goes out of map boundaries. Woohoo to that.
General fixes all around. Two new assets were made in preparation for a
test game (IWJTest.java).